--- commands/integration_commands.py
@@ -1,73 +1,0 @@
-"""Quick integration commands for testing subsystems together.
-
-Generated by Claude for quick bring-up testing.
-
-Bindings (see data/inputs/2026bot.yaml):
-    Operator A      - toggle intake deploy / retract
-    Operator B held - run intake rollers
-    Operator X      - toggle flywheel 6000 RPM + hood 0 deg
-    Driver LB held  - hopper 50% + feed 30%
-"""
-
-import commands2
-
-from constants.swerve_constants import HopperConstants
-
-
-# -- Intake toggle deploy/retract -------------------------------------------
-
-def toggle_intake_deploy(intake):
-    """Return a command that deploys if stowed, or stows if deployed."""
-    def _toggle():
-        if intake.isIntakeDeployed():
-            intake.stowIntake()
-        else:
-            intake.deployIntake()
-
-    return commands2.cmd.runOnce(_toggle, intake)
-
-
-# -- Flywheel spinup toggle --------------------------------------------------
-
-_SPINUP_RPM = 6000
-_SPINUP_HOOD_ANGLE = 0.0
-
-
-def toggle_spinup(shooter, hood):
-    """Toggle flywheel on at 6000 RPM with hood at 0 degrees.
-
-    Uses shooter.flywheelActive as the toggle state. When toggling off,
-    RPM is set back to 0.
-    """
-    def _toggle():
-        shooter.toggleFlywheelActive()
-        if shooter.flywheelActive:
-            shooter.setRPM(_SPINUP_RPM)
-            hood.setAngleDegrees(_SPINUP_HOOD_ANGLE)
-        else:
-            shooter.setRPM(0)
-
-    return commands2.cmd.runOnce(_toggle, shooter, hood)
-
-
-# -- Hopper + feed while held ------------------------------------------------
-
-_HOPPER_POWER = HopperConstants.defaultPower
-_FEED_DUTY_CYCLE = 0.3
-
-
-def run_hopper_and_feed(hopper, feed):
-    """Return a command that runs hopper at 50% and feed at 30% power.
-
-    The command runs continuously while the button is held, then cleans up
-    on end (interrupted or cancelled).
-    """
-    def _start():
-        hopper.setPower(_HOPPER_POWER)
-        feed.setPower(_FEED_DUTY_CYCLE)
-
-    def _end():
-        hopper.stop()
-        feed.stop()
-
-    return commands2.cmd.startEnd(_start, _end, hopper, feed)
